requirements.txt
fastapi==0.110.0
loguru==0.7.2
SQLAlchemy==2.0.29
alembic==1.13.1
asyncpg==0.29.0
greenlet==3.0.3
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
uvicorn==0.29.0
pytest==8.1.1
httpx==0.27.0
pytest-asyncio==0.23.6
black==23.9.1
isort==6.0.0b2
flake8==7.0.0
mypy==1.9.0
starlette==0.36.3
pydantic==2.6.3